+++ tests/supertest.js
@@ -1,14 +1,27 @@
 const request = require('supertest');
 const app = require('./../app');
 const assert = require('chai').assert;
+const util = require('util');
 
 describe('Supertest', function() {
 
   describe('POST /slack/streak', function() {
     it('returns playername and streak number', function(done) {
-      let uniquePlayerName = '@UniquePlayerName' + Math.floor((1 + Math.random()) * 0x10000);
-      let message = 'text=@Laura and ' + uniquePlayerName + ' crushed @Peter + @Mary today, winning 10:0';
-      let expected = '@Laura and ' + uniquePlayerName + ' played 10:0 against @Peter and @Mary.';
+      const uniquePlayerName = slackify();
+      const message = util.format('text=%s and %s crushed %s and %s today, winning %d:%d',
+        slackify('Laura'),
+        uniquePlayerName,
+        slackify('Tom'),
+        slackify('Mary'),
+        10, 0);
+
+      const expected = util.format('%s and %s played %d:%d against %s and %s.',
+        slackify('Laura'),
+        uniquePlayerName,
+        10, 0,
+        slackify('Tom'),
+        slackify('Mary')
+      )
       request(app)
         .post('/slack/result')
         .set('Content-Type', 'application/x-www-form-urlencoded')
@@ -27,7 +40,7 @@ describe('Supertest', function() {
               }
 
               assert.isOk(res.text);
-              assert.match(res.text, /^@UniquePlayerName[a-zA-Z0-9]+ has won [0-9]+ game\(s\) in a row.*/);
+              assert.match(res.text, /^<@[a-zA-Z0-9]+|@UniquePlayerName[0-9a-zA-Z]+> has won [0-9]+ game\(s\) in a row.*/);
               done();
             });
 
@@ -35,23 +48,7 @@ describe('Supertest', function() {
     });
   });
 
-
-
-
-
-  describe('POST /slack/result', function() {
-    it('returns 201 on success', function(done) {
-      let uniquePlayerName = 'UniquePlayerName' + Math.floor((1 + Math.random()) * 0x10000);
-      let message = 'text=@Laura and @' + uniquePlayerName + ' crushed @Peter + @Mary today, winning 10:0';
-      request(app)
-        .post('/slack/result')
-        .set('Content-Type', 'application/x-www-form-urlencoded')
-        .send(message)
-        .expect(201, '@Laura and @' + uniquePlayerName + ' played 10:0 against @Peter and @Mary.')
-        .end(done);
-    });
-
-    it('returns 400 on missing body.text', function(done) {
+  it('returns 400 on missing body.text', function(done) {
       let message =  'something=@Laura and @Dennis crushed @Peter + @Mary today, winning 10:0';
       request(app)
         .post('/slack/result')
@@ -62,7 +59,12 @@ describe('Supertest', function() {
     });
 
     it('returns 400 if body.text can not be parsed (missing goals)', function(done) {
-      let message =  'text=@Laura and @Dennis crushed @Peter + @Mary today, winning';
+      const message = util.format('text=%s and %s crushed %s and %s today',
+        slackify('Laura'),
+        slackify('Ben'),
+        slackify('Tom'),
+        slackify('Mary')
+      );
       request(app)
         .post('/slack/result')
         .set('Content-Type', 'application/x-www-form-urlencoded')
@@ -72,7 +74,7 @@ describe('Supertest', function() {
     });
 
     it('returns 400 if body.text can not be parsed (missing players)', function(done) {
-      let message =  'text=Laura and Dennis crushed Peter + Mary today, winning 10:3';
+      const message = 'text=Somebody played today 1 : 10';
       request(app)
         .post('/slack/result')
         .set('Content-Type', 'application/x-www-form-urlencoded')
@@ -80,5 +82,16 @@ describe('Supertest', function() {
         .expect(400)
         .end(done);
     });
-  });
 });
+
+/**
+ * 
+ * @param {*} name if omitted, we generate something unique
+ */
+const slackify = function(name) {
+  if (!name) {
+    return '<@' + Math.floor((1 + Math.random()) * 0x10000) + 'AAAAAA|UniquePlayerName' + Math.floor((1 + Math.random()) * 0x10000) + '>';
+  } else {
+    return '<@01234FFFFFF|' + name + '>';
+  }
+}
